requirements.txt
requests
colorama
discord.py
